requirements.txt
streamlit
opencv-python-headless
deepface
numpy
Pillow
